# swagger.yaml
openapi: 3.0.0
servers:
  - url: http://127.0.0.1:8080/
    description: path to the main simpleca API
info:
  description: Simple certificate authority web server
  version: "0.0.1"
  title: "Simpleca API"
  contact:
    email: cyd@piratemind.com
  license:
    name: MIT
    url: 'https://opensource.org/licenses/mit-license.php'
paths:
...
  /csr:
    post:
      summary: Generate a certificate signing request from a private key
      operationId: generateCSR
      parameters:
        - in: query
          name: passphrase
          required: false
          allowEmptyValue: true
          description: passphrase of the private key
          schema:
            type: string
        - in: query
          name: C
          required: false
          allowEmptyValue: false
          description: Country
          schema:
            type: string
          example: FR
        - in: query
          name: ST
          required: false
          allowEmptyValue: false
          description: State
          schema:
            type: string
          example: France
        - in: query
          name: L
          required: false
          allowEmptyValue: false
          description: Locality
          schema:
            type: string
          example: Paris
        - in: query
          name: O
          required: false
          allowEmptyValue: false
          description: Organization
          schema:
            type: string
          example: MyOrg
        - in: query
          name: OU
          required: false
          allowEmptyValue: false
          description: Organization unit
          schema:
            type: string
          example: MyUnit
        - in: query
          name: CN
          required: true
          allowEmptyValue: false
          description: Common name
          schema:
            type: string
          example: localhost
        - in: query
          name: altnames
          required: false
          allowEmptyValue: false
          description: "Alternate names (coma separated list)"
          schema:
            type: string
          example: localhost
        - in: query
          name: ips
          required: false
          allowEmptyValue: false
          description: "IP addresses (coma separated list)"
          schema:
            type: string
          example: 127.0.0.1
      requestBody:
        description: The private key
        required: true
        content:
          octet-stream:
            schema:
              type: string
            examples:
              keyoctet:
                $ref: '#/components/examples/keyoctet'
      responses:
        '200':
          description: here is the certificate signing request
          content:
            text/plain:
              schema:
                type: string
              examples:
                csrtext:
                  $ref: '#/components/examples/csrtext'
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/output'
              examples:
                csrjson:
                  $ref: '#/components/examples/csrjson'
        '400':
          description: bad input parameter
          content:
            text/plain:
              schema:
                type: string
              examples:
                converterror:
                  value: Unable to convert to private key
                keytype:
                  value: Wrong key type
                wrongcn:
                  value: Common name can not be empty
        '405':
          description: not a valid method
        '500':
          description: internal server error
          content:
            text/plain:
              schema: 
                type: string
              examples:
                gen:
                  value: Error while generate certificate signing request
...
components:
  schemas:
...
    output:
      type: object
      properties:
        key: 
          type: object
          properties:
            priv:
              type: string
              readOnly: true
            pub:
              type: string
              readOnly: true
        crt:
          type: string
          readOnly: true
        csr:
          type: string
          readOnly: true
        ca:
          type: object
          properties:
            key:
              type: string
            crt:
              type: string
  examples:
    keyoctet:
      summary: example of private key in octet stream
      value: |
        -----BEGIN RSA PRIVATE KEY-----
        MIIEogIBAAKCAQEAvrczWwufabnJelO1/3tOe7gxAVaH4LHUg3KLq4xtJLvzxZ2g
        S89gZ8jsBgh7o57CoGT8UM2TT5YivQNpiiCV4VNJr2eUZTS22L/GchQR3wWR/cqd
        mt9bMG/vHK9CqHgxTsZLf+nABEWDJypgpLRbLJJYLprhZrn566+CQ6c+UbFVaPln
        wUydzetUW2yd5X91G16DPgLlMZejD05PbdUJhjgF7dj6sU+nnSpvLP69x9gPQNi8
        4vWohZH5LMNYxftHWfVE9oWTabuJwK8iE2UCnPoK4z6LW10PD5A9GMcLq3EAvVPm
        R6ZgxJNoY7ldC0b/d2XXLvlmbk8QHDxwak0hiQIDAQABAoIBABZVqBsCenHMQy8j
        PI4l56QPhxWsB4+l48hdEsRNVr7VkA/VjN8z8wtUeWefeLNJtWtJSE9CM/0/DORr
        0OFPPZ48vcVf8P4FNWCe3OOhPBKNOXKqVti7TZeXlD4cvEa0hLm/+TmKq2UUqu5c
        RSMkyASpQtB6bOk14HHchPndTSwNz7PmpQgE0DtvNmKhnTP/effNgo3XbYDvJbd4
        kd3x+HzSgXnNZaEB5nFfkUC+HI9gqpz/h+e1rZr3sK4r3aIw7Kp+d4azyTkM+IZ6
        SvZCZO8HsQfCwq65JPntNprGse5d6Z0v02rm/aqWELmb0PUCZX+4tNW8B23SX3zS
        ic9QVbECgYEA1PpGKhMRlfJixQdF8e2obrpj67H+i9idMRJFz3gjgZERcgrC+MwU
        Ls2owtoo+c/UlO/2mQn32mrLH1IoDmj2wdjwWp2tOTJ/NX7x90dOG2RQxPhZ+NE+
        3SEsKoXxBB3ij6Kw670bb0q4l7D5eCj/zJHxU/Ovawh1MrnbVILu4X0CgYEA5T2w
        pDmo7V2JXc0622l1OqMKDovIu2Y7XaeRaTpPa2EWfZh5l1iOcbrtFr/LNxgX9UYa
        2Mpt3hfuhZqL9CN+tK/6ipQ9dDz3XXKkvJLfqSqyrV6s2s3Ei/RBTIL7fzqTAxiF
        RcQRC4HNXN31rUlu+XR5h9puMz07LzsnJnLrvf0CgYBZNwG/TD+YnXFscZCGOHTO
        9Xs311b2w+WNEcd2ki/lbZxuNj3YNLzbokDoZrS1eK7nOOIG15sgvo2Z8DNhXezH
        5RVCEDzo8dNB2IyJdjAs30fKSvJ9fUGlJek3YuZEQDSQ0KjmB1B/vUA4IuVNxXbp
        /dSmwzSfVGqrthk/fJ2OjQKBgD29vryMdw3vGlc0NjzwC03OqZ+/UEIAx85ymjxg
        IxHFpVwQW/dCN6Mz+DUfVOkyyQkbZ5Gp+6tMgMZlGHqZsL+TggQe5upOnUUC4pGT
        cT/LqzH8hMutjdkv4QJtaLUcnom/6e0fYIPb4ua3NTsFwyjWkQDg+Ls+K5mN8f5u
        B0Y1AoGAPhkvKDWW8YeerkKcN3wJbOHibOphmg68feXCvkrrkyoSuDqt39R0ZR7z
        3lc8uV8uPuZPFRJXdaVerxykXyfer+9zroSUsGMBuLMVviPO80cIZTeCmjZb1egz
        a17Pd/TCZgT9yjr7OvJOjyCXN3hYNeKK/C1msMgmIxx8lc6agMM=
        -----END RSA PRIVATE KEY-----
...
    csrtext:
      summary: example of certificate signing request in plain text
      value: |
        -----BEGIN CERTIFICATE REQUEST-----
        MIIC9zCCAd8CAQAwYzELMAkGA1UEBhMCRlIxDzANBgNVBAgTBkZyYW5jZTEOMAwG
        A1UEBxMFUGFyaXMxDjAMBgNVBAoTBU15T3JnMQ8wDQYDVQQLEwZNeVVuaXQxEjAQ
        BgNVBAMTCWxvY2FsaG9zdDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEB
        ANCCLXbfA9DRXb4Iq+1ugGjaW20Tcc5/GCYcbVuPjWPgx5ZcKyxhCQndIS05mRAT
        aNPJlyPUGiry7emx/bckzmg23Ok1ZlbO6SlWrvRlFxR/PzTXK0xLe8uASsMC9J4V
        O2B1x7oVS7P7E9J2ZKcdqppnsOWCGYMxWv9kpcALFh1s4jumBWcDCephNtF+AtZn
        BE+lJDDAQtqaiT3tjROOYFVoMShAyxEH3agUuGOI/A0kW5mlhr7rzD9bVVi/J+Y9
        Av5yDsuL1Gf5YEHr7dUo78GabpI9HuxgLxxL9qt3I6b7P4DTHmIiRCVF2YtUILiw
        Id/96olP6QvTotR3SUbblcECAwEAAaBPME0GCSqGSIb3DQEJDjFAMD4wPAYDVR0R
        BDUwM4IRd3d3LmxvY2FsaG9zdC5jb22CBHRlc3SCEnRlc3QubG9jYWxob3N0LmNv
        bYcEfwAAATANBgkqhkiG9w0BAQ0FAAOCAQEA0IIiMvj/GjKhxIwRucTtEMsuQRCe
        l5NhIympe8odDkd39NHXEreqwc1Xxz2rsDq4f4dLap8bJqOnU1y53LJhlcRfrMcC
        s0/zDZ5yQtsP1ENhr/Rb8BkvMju0WUD2nfpSTKlPirBB8mK63iq1EXR/gRFx5J37
        7rxg9cf+plKuA2a/lVxYd9E8QENWqj2LmWS5y4OnageGHz2WDZJDM+ov2S68eb4q
        9JrxHlqJk7kRWBYBmSjO+66vOs4Kl0KcdepJja5FCnebUeXWusC0dgtxeR1xWE8q
        AHNyaY33J1wVIcjV0Vb82BYnaaVNLiemEfx3fW430tSDH7453mm9Xo3VyQ==
        -----END CERTIFICATE REQUEST-----
    csrjson:
      summary: example of certificate signing request in json
      value:
        csr: |
          -----BEGIN CERTIFICATE REQUEST-----
          MIIC9zCCAd8CAQAwYzELMAkGA1UEBhMCRlIxDzANBgNVBAgTBkZyYW5jZTEOMAwG
          A1UEBxMFUGFyaXMxDjAMBgNVBAoTBU15T3JnMQ8wDQYDVQQLEwZNeVVuaXQxEjAQ
          BgNVBAMTCWxvY2FsaG9zdDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEB
          ANCCLXbfA9DRXb4Iq+1ugGjaW20Tcc5/GCYcbVuPjWPgx5ZcKyxhCQndIS05mRAT
          aNPJlyPUGiry7emx/bckzmg23Ok1ZlbO6SlWrvRlFxR/PzTXK0xLe8uASsMC9J4V
          O2B1x7oVS7P7E9J2ZKcdqppnsOWCGYMxWv9kpcALFh1s4jumBWcDCephNtF+AtZn
          BE+lJDDAQtqaiT3tjROOYFVoMShAyxEH3agUuGOI/A0kW5mlhr7rzD9bVVi/J+Y9
          Av5yDsuL1Gf5YEHr7dUo78GabpI9HuxgLxxL9qt3I6b7P4DTHmIiRCVF2YtUILiw
          Id/96olP6QvTotR3SUbblcECAwEAAaBPME0GCSqGSIb3DQEJDjFAMD4wPAYDVR0R
          BDUwM4IRd3d3LmxvY2FsaG9zdC5jb22CBHRlc3SCEnRlc3QubG9jYWxob3N0LmNv
          bYcEfwAAATANBgkqhkiG9w0BAQ0FAAOCAQEA0IIiMvj/GjKhxIwRucTtEMsuQRCe
          l5NhIympe8odDkd39NHXEreqwc1Xxz2rsDq4f4dLap8bJqOnU1y53LJhlcRfrMcC
          s0/zDZ5yQtsP1ENhr/Rb8BkvMju0WUD2nfpSTKlPirBB8mK63iq1EXR/gRFx5J37
          7rxg9cf+plKuA2a/lVxYd9E8QENWqj2LmWS5y4OnageGHz2WDZJDM+ov2S68eb4q
          9JrxHlqJk7kRWBYBmSjO+66vOs4Kl0KcdepJja5FCnebUeXWusC0dgtxeR1xWE8q
          AHNyaY33J1wVIcjV0Vb82BYnaaVNLiemEfx3fW430tSDH7453mm9Xo3VyQ==
          -----END CERTIFICATE REQUEST-----
...